Modified data_model_to_mdl to export Scene.geometry.mdl as dynamic.
To make the MDL files more consistent for hand editing, this version
exports a geometry file whether dynamic geometry is used or not. The
previous version would include the DYNAMIC_GEOMETRY directive in another
file, but this version includes that file in "Scene.geometry.mdl" when
exporting modular MDL files. This makes Scene.geometry.mdl the file to
look at whether exporting static or dynamic geometry.

With this change, the Scene.main.mdl file only includes other files and
sets the seed.