Printed frame numbers mod 100 when generating dynamic geometry.
Generating dynamic geometry can take some time, so this commit just
gives some reasonable feedback that something is happening.

The generation time may be worse when static objects are mixed in with
dynamic since MCell requires that all objects have dynamic data when
any are dynamic. The exporter could recognize this situation and only
generate a single object that's included multiple times to meet MCell's
requirement, but the current version doesn't do that. There is a check
whether obj['dynamic'] is set, but that's currently over-ridden to
output dynamic geometry for all objects.