Added support for optional scene name in data_model_to_mdl.write_mdl.
This commit removed the "hard-coded" name of "Scene" from the data model
exporting of data_model_to_mdl.py. The main function (write_mdl) now
takes a scene name as a parameter with a default of "Scene". There may
likely be other places in CellBlender that still use a hard-coded Scene
name, so the code may not work with any name yet. But this is a start.