Added Export Scripting Support to data_model_to_mdl.py
This commit supports exporting of both raw MDL and MDL written by a
Python script. This capability had existed in the io_mesh_mcell_mdl
code, but was missing in the data model exporter (data_model_to_mdl).

Because the exporting is not identical to the io_mesh_mcell_mdl code,
it should be carefully tested to ensure that it's doing what we want.