Added an "ignore_cellblender_data" flag and removed show options.
An "ignore_cellblender_data" flag was added to the scripting panel
to allow ALL of the exported MDL to come from scripts. This allows
CellBlender to be used for pure MDL scripting without interference
from any of CellBlender's data. This is a somewhat experimental
feature, and certain display items (like molecule properties) will
not show up with pure MDL. But it may prove useful for teaching,
and it wasn't very hard to do.