Added return values indicating whether MDL was actually written or not.
The MCell INCLUDE_FILE directive fails when the included file is either
empty or contains only comments.

The current data model export code delegates the writing of MDL to
functions for different subsections of the model. Those functions may or
may not actually write anything (depending on what's in their portion of
the data model). This is not a problem when writing a unified MDL file
because the file is guaranteed to not be empty because certain sections
are always written. However, when exporting modular MDL, it is possible
that a given section of the data model will not generate any MDL. In
that case, including the empty file (even if a comment were written)
will cause MCell to fail.

This commit continues to delegate the MDL writing to functions, but the
functions are now required to return a Boolean flag indicating whether
they have actually written meaningful (non-comment) MDL to their file.
This flag is then used to determine whether to write an INCLUDE_FILE
directive into the main file for that delegated section (or not).

Note that this mechanism may generate empty files when exporting as
modular MDL, but those files will not be referenced in the main MDL
file.