Fixed logic bug in recent commit to handle empty MDL exports.
The strategy for exporting without empty MDL files (while still
delegating) has the delegated function return whether any MDL was
actually written (or not) as a boolean flag. If none was written,
then the file won't be included in the parent (Scene.main.mdl).

The static geometry and release sites share the same file (named
"Scene.instantiation.mdl") so they may both contribute to the need
to include that file. This was handled with an "or" of the two
results. However, the second result was being returned by a function
which was not being called when the first value was True (optimization).
The offending line was:

  actually_wrote = actually_wrote or write_release_sites(...)

The first "actually_wrote" came from writing the static geometry,
and it was being "or-ed" with the result of "write_release_sites().

This has now been separated to work properly.

More testing is still needed on this version, but it's getting close.