Initial working (limping) version of sweeping scripted dynamic geometry
In order to sweep dynamic geometry, the dynamic geometry generation
function needs a copy of the data model (or at least the parameters from
the data model).

There are a couple of problems with this. First, the current frame
change handler doesn't have a data model to pass to the dynamic geometry
generation function when moving the time line. Second, the parameter
expressions are strings which will need to be evaluated before they can
be used for dynamic geometry modification.

The first problem is currently handled by selecting the "Files" display
option for each dynamic object (which will display from the generated MDL
rather than by calling the script itself). Hopefully, this problem can
be better solved by maintaining an updated data model. Since that wasn't
the original purpose of the data model, it may require some thought. For
now, we could just update the data model upon starting a run or when the
user requests an update manually.

The second problem just means that the user script has to deal with a
string expression rather than a numeric value. We do have a number of
functions in the parameter system code to manage expressions, but they
operate in our "Blender parameter system" environment rather than the
"data model" environment. In practice, however, sweeps should be defined
as purely numeric values since the sweeping code also doesn't know how
to deal with parameter expressions without resorting to the "Blender
parameter system".