Made data model available to the frame change handler for sweeping
The previous version was able to pass the data model to the user's
dynamic geometry script when generating the data, but the frame change
handler did not have a data model available to pass to the same dynamic
geometry script when moving the time cursor. So the same dynamic
geometry was drawn regardless of the point in the sweep space. Note that
this only happened when drawing from the script. If the drawing was done
from the MDL at each sweep point, then the geometry was correct.

This commit adds a "cellblender.current_data_model" field that is either
"None" or the data model that was most recently run. This data model can
be used (along with the mol_viz "choices_list") to draw the proper
geometry for each frame consistent with the parameter settings being
viewed (currently in the "mol_viz" panel).

The current code can draw dynamic geometry from 3 sources:

  1. A script
  2. The dynamic geometry MDL files
  3. Other (doesn't actually draw at all) typically shape keys