Added some dynamic geometry features to assist in script writing
 - Call mcell.model_objects.update_scene when refreshing molecules
 - Wrap try/except around opening the data_layout.json file

The first feature makes it easier to change a script and just hit the
refresh button to see the result. This previously required changing the
frame to force the frame change handler to redraw.

When running swept dynamic geometry, the MDL files describing the
geometry might be in subdirectories based on sweep position. However, if
the model hasn't been run at all, there will be no files. This produced
lots of file-not-found error messages. The try/except eliminates these.

Note that there can be a disconnect between what CellBlender knows about
the sweeping and the current output from an MCell run. This has always
been an issue, but it can be more noticable when the user is changing
dynamic geometry scripts. These changes try to make it easier.