Added an "output" entry to each task_dict[pid] to hold output lines
In order to display the simulation output without using bl_text, each
task now contains a list of lines that's dynamically updated using the
same mechanism used to update the bl_text. The list is stored in:

  self.task_dict[pid]['output']

This output list is filled by the "write_output" function when it is
passed a non-None value for its "output_list" parameter. This is similar
to the current handling of the bl_text parameter.

The percentage done timer in cellblender_simulation.py was also updated
to use this new list rather than the bl_text (the old code was commented
out rather than removed in this commit).

This now produces percentage complete updates even when not saving
blender text logs. Note that text logs can still be saved, but they are
not used to calculate the percentage complete displayed in the run
panel.