Added minimal support for overlay text showing results of each run.
This is a crude implementation of overlay text borrowed from one of the
runners (sim_runners/queue_local/__init__.py).

Because this code makes a call to "SpaceView3D.draw_handler_add", it is
currently activated through an operator which must be called to enable
this mode (essentially add the handler). So the "Show Text" button must
be pressed one time after starting Blender to enable it. It will remain
enabled until Blender exits (or it is manually disabled via "Hide Text").

This commit modifies the sim_runner_queue.py file to include an "output"
key in the task dictionary. This will hold a list of lines and is passed
via the "output_list" parameter to both "write_output" and "run_proc".
The list currently holds both stdout and stderr with no distinction
between them. It might be possible to tag them with their source, but
it's a bit complicated because the "read_output" function just gets a
pipe without knowing whether it's stdout or stderr.

This commit retains the direct writing to Blender's bl_text, but it is
disabled by default. We might optionally remove this and simply provide
a mechanism to copy from the internal text buffer to either Blender
text(s) or to plain file(s).

This version still lacks a number of nicer features that were included
in the Queue Runner version (selecting number of lines, changing the
update time interval), but these can be added along with other cosmetic
changes to the interface.