Added an ISSUES.txt file for tracking issues across Git Environments.
The ISSUES.txt file is intended as a way for CellBlender developers to
document and communicate issues that may need attention, and to document
when those issues have been resolved.

The format (as plain text) is pretty much free form, so any notes are
welcome. It's probably good to have an "Open" and a "Resolved" section
to make it easier to see what needs to be done. But different layouts
are welcome (it's all backed up in the history anyway).