Fixed a few logic problems related to static scripted objects.
Static scripted objects never get their data from "Exported Files"
because they don't export their goemetry in the dynamic format (the
purpose of the "Exported Files" option). So the ability to choose the
"Exported Files" has been removed when using static scripted geometry
for an object.

Similarly, when a scripted object is NOT dynamic, it must get its
geometry from the script (and not from a generated file).

This commit also adds a "dynamic" boolean to the user's script
environment. This enables a geometry script to detect whether it
should be generating dynamic geometry or static geometry. So a
dynamic script may generate a range of sizes when used as dynamic
geometry, but generate only selected sizes when used as static
geometry. For static cases, the frame_number is also set to -1,
but the more reliable signal should be the dynamic flag itself.