Renamed "current_data_model" to "browser_data_model" in data_model.py
The recent addition of "current_data_model" to cellblender itself leads
to potential confusion when searching (grep) for "current_data_model".

This commit renames the local variable "current_data_model" inside the
data_model.py file to be "browser_data_model" (since it is created for
browsing the data model) to avoid confusion.