Rebuilding of Data Model to support dynamic geometry scripts
The data model was originally designed only to store Blender's
property database in a "non-volatile" form to allow upgrading
of older models in newer versions of CellBlender. In that role,
it was only generated when saving the .blend file.

Over time, the data model has taken on a greater role which
added requirements for it to be updated (when exporting MDL,
for example). With the addition of scripted static geometry,
the data model should be updated whenever the static geometry
is being redrawn so the script getting the data model will
have the latest parameter settings.

These new demands suggest that an updated data model should
be kept on hand at all times. However, that would require a
lot of callbacks added to every property to update the data
model whenever a property is changed. As a compromise, the
data model (or just parts of the data model) may be updated
only at certain times.

This commit adds updating of the parameter system part of
the "cellblender.current_data_model" whenever the "Refresh"
button is pressed. If the current_data_model is None, then
the entire data model will be regenerated.

This is a temporary measure to (hopefully) meet the update
needs of the static scripted geometry system. In the long
run, it might be better to maintain a current data model
at all times and let that be the "one true model" within
CellBlender. In that case, the Blender properties that
currently hold most of the model would be reduced to being
just user interface items for updating the data model. The
advantages in speed might be significant since Blender's
property system performance degrades as the number of
properties grows large.