Partially working MCellR with MCell Local engine/runner.
This version was able to complete several of the MCellR models and
produce what appear to be reasonable results.

There is still a problem with sweeping and dynamic geometry when run
with MCellR. That needs more investigation.

This version scans the molecule definitions to automatically detect
whether the data model is MCell or MCellR.