Improved plotting swept data with a ppt= parameter to plotters
The ppt=<string> parameter is passed to plotting functions where the
string indicates the point in parameter space ("x=4.0,num=8"). Any
plotting functions written to look for that parameter may use it as
appropriate (typically for labeling). Currently, most plotting functions
use the file name parameter (f=) for labeling, but for parameter sweeps,
the file name can produce labels that are long and difficult to read.
The new "ppt" option stands for "parameter point" since it represents a
point in the swept parameter space.

Since all plotting interfaces should be designed to ignore parameters
they don't understand, this should have no effect on any plotting
interfaces that haven't been updated to look for the new parameter.
This commit includes updates to the MPL, Java, and Simple plotters to
accept and process this new parameter.

Note that both the "ppt=" and "f=" parameters will be sent to all
plotting functions. The choice as to which to display is up to each
plotting function. Right now, the MPL and Simple plotters use the new
ppt option, while the Java plotter has a menu option to show the file
name if desired. That level of user interface is up to the designer of
each plotter. We could make that option available in the Plotting Panel,
but that might be better done during an overhaul of the plotting code.