Added "exist_ok=True" to os.makedirs calls, fixed swept project path
Added the "exist_ok=True" to os.makedirs calls while tracking down the
source of the double "output_data" directory for MCellR models.

The problem was in the "write_mdlr" function which was making improper
assumptions about the file_name parameter passed to it. That file_name
is the final location of the MDL with the "output_data" portion already
appended along with each of the swept subdirectories.

It's not clear whether the "exist_ok=True" parameter is needed, but it's
not a bad idea to leave it there in case it should be needed.

This version can run parameter sweeps on MCellR models.