Added a new plotting issue found during testing of MCellR support.
The plotting panel contains a setting to generate all plots (ignoring
the current visibility setting) or to only generate the plots selected
as visible. This was added long ago to cut down on the amount of data
that gets plotted, but it might not be in the data model or be honored
by data_model_to_mdl.py.

This is a relatively low priority issue, but worth having on a list.