Changed the release site id logic to check for name collisions.
The previous logic worked when all sites were added through the user
interface. However, when sites can be imported from the data model,
those additions did not update the next_id. This could cause name
collisions because the next_id didn't know that those previous names
(and associated ids) were in use.

This name collision could also be induced by manually changing a site
name to an anticipated auto-generated name. For example:

  1. Create a release site (it will default to "Release_Site_1")
  2. Rename it to be "Release_Site_3"
  3. Create two more release sites ... the second will conflict

The current fix adds a little bit of processing time via the list
comprehension to obtain the list of all names and to check that the
proposed new name isn't in the list. In the typical case, that's the
total cost. In very odd cases, additional processing may be required
when the proposed name should be the first in a long list of sequential
names that had somehow been constructed using "Release_Site_#" names.

Note that this change should be made to all similar naming constructs.