Attempted to move writing parameters before initialization for MCellR
This change moves the writing of the parameter system before writing the
initialization (which includes ITERATIONS, TIME_STEP, AND VACANCY...).
This move follows the normal MCell MDL pattern of parameters before any
MDL that might use parameters. However, this seemed to cause MCellR to
not run (at least sometimes).

So while this *SHOULD* be the proper export sequence, it may need to be
changed to get MCellR to work properly.