Write parameterized ITERATIONS, TIME_STEP, and VAC_SRCH_DIST first
This commit converts expressions for these keywords to write first:

  ITERATIONS
  TIME_STEP
  VACANCY_SEARCH_DISTANCE

This was done to fix errors reported by MCellR when parameters were
included first (which should be the proper order to export). This also
more closely matches the MDLR format given as examples.

However, additional testing indicated that MCellR would sometimes run
fine with the parameters placed before these three keyword values. That
testing also indicates that this fix doesn't guarantee a working model.
MCellR still seems to fail to run sometimes.

So it's not clear whether or not this commit is needed. It is being
added for now to preserve the conversion work.