Defined major property groups on a single line for easier searching.
CellBlender is made up of a number of property groups that are assigned
to variable names in cellblender_main. When working with code and
looking at a particular property group, it's often helpful to quickly
find the name associated with that group. However, since the former
layout of these groups put the group name on one line, and the assigned
name on another line, a simple "grep" would not show both together.

This change puts the group name on the same line as the assigned
variable name to make this easier. It also makes the code more compact.