Changes to support running on Sun Grid Engine (SGE)
Added try/except to sections of the code that rely on Blender either
directly or indirectly (through CellBlender).

Both data_model_to_mdl.py and data_model.py were designed to be able to
operate outside of Blender (and without CellBlender). However, in the
rush to get MCellR working, several functions were imported that
violated that design goal. This caused them to no longer work within the
SGE running environment. This commit simply adds try/except blocks
around code that might otherwise fail without the required imports.

At the time of this commit message, these changes have worked to submit
jobs to the local SGE host, but the jobs are still waiting because the
queue is so full. So it's not yet known if they will complete or not.