Changes related to fixing CellBlender's CPU leak
  - Included tasks with status of "died" or "mcell_error" when
    counting completed tasks to stop the percentage_done_timer.

  - Reduced the number of percentage_done_timer starts.

  - Added an optional "heart beat" to print on timer ticks.

Running "top" was showing that under certain conditions, CellBlender
would gradually increase its CPU usage (typically after many runs).
This was noticed to be more common when running MCellR projects.
The CPU usage was cumulative - increasing with additional runs as
long as Blender was running.

It turns out that the counting of completed tasks was based on the
task's reports of iteration number. So any tasks that had died were
considered still active. This kept the percentage done timers running
in the background. This was more common (and noticable) with MCellR
runs which terminate badly at a higher rate for unrelated reasons.

This commit checks for tasks with status of "died" or "mcell_error" to
include them in the count of completed tasks. This might be the proper
thing to do for all tasks rather than checking iterations.

This commit also added an optional "Print Timer Ticks" boolean in the
"Settings & Preferences" panel (defaults to off). This provides a way
to verify whether the percentage_done_timer is running or not. This
was helpful for debugging and may be useful in the future.

Finally, the previous version was starting a percentage_done_timer
for each mcell run. This commit only starts one timer for the batch
of runs (since the timer loops through all processes in the list).
It may also be worth checking whether that timer is running first,
but it's not clear how to get at that in CellBlender.