Set first new molecule automatic name index to "1" after all are deleted.
Molecules get an automatic name based on the "last_id" counter. This id
is incremented after creation of each new molecule, and it is only reset
to 0 when there are no molecules (this follows the pattern for other
items in CellBlender.

The "last_id" counter was properly managed in the allocate_available_id
function, but it was being reset to 1 (rather than 0) in two other
locations. The symptom was that the first new molecule would be numbered
"2" rather than "1". This commit resets "last_id" to 0 in those other
locations so that new molecules (after all have been deleted) will start
with 1.