Changed "Show All/Multiple" label to "Multiple Panels"
The original "Show All/Multiple" button was intended to toggle between
three states (showing one, showing multiple, showing all). But the
showing all was problematic for several reasons and with so many panels
was of little practical use. So it's functionality is now to show either
one panel at a time or allow selections of multiple panels at once. So
it is better described as "the Multiple Panels button". This commit
makes that change.

Note that this is also in conjunction with an effort to update all of
the main button panels in the tutorials. Right now, each tutorial has a
copy of its own top-level button panels. These can be inconsistent and
also out of date. To standardize presentation, the "shared" directory in
the tutorials section will have one copy of the main button panel with
each button selected. The tutorials can then reference these shared images
rather than creating their own. When the main panel changes, they can be
changed in one place and that will update all of the tutorials using them.