Brought back ParameterSpace.py from 0baf0fed0d4e370a0fd9b93b24bfce2a69fc0e5e
The Parameter System has been implemented using a number of technologies
within Blender. The original version was pure Python and it did not
depend on any Blender modules or Blender callbacks for its functionality.

Over time, Blender's own update callback system was used as a more natural
mechanism for the propagation of updates. However, this has proved to be a
very slow mechanism with a number of other complications.

This commit brings back ParameterSystem.py from April 4th, 2014 to use as
an experiment. It is not being added to the distribution at this time, but
it is being committed so changes can be tracked.

This version was found with:

  git rev-list -n 1 HEAD -- ParameterSpace.py

That returned commit: 02106a707768d6eac96092293b0b4a2159f6a01e

It's parent commit was: 0baf0fed0d4e370a0fd9b93b24bfce2a69fc0e5e

The file was restored with this command:

  git checkout 0baf0fed0d4e370a0fd9b93b24bfce2a69fc0e5e -- ParameterSpace.py

This commit adds it to the current development branch.