Improved profiling and efficiency of ParameterSpace.py
This code can update parameter values very quickly, but it can take a
while to do the initial parsing. Again, this is pure Python code and is
not yet being used in CellBlender. It's being updated here to track its
history from the original 2014 version in the hope that it might replace
the current parameter code.