Added automatic update of forward reference corrections
It is best if parameters are defined based on existing parameters.
Otherwise they will have an unknown value and the parameter system
will have to continually search for resolution as each new parameter
is added.

This commit handles the case where that policy is not observed and
parameters are defined that depend on undefined parameters. In that
case, the parameter system will attempt to resolve undefined forward
references whenever a parameter name is changed. This can be a more
time consuming process with many parameters, so it's best to define
them in the proper order.