Added test_suite/tutorial_maker.py from simple cone creation demo.
This is the first step in a project to build tutorials from code.

The code would be very similar to our existing test suite. It would
invoke operators just as if they'd been invoked from the user interface.
It would be scripted to carry out a particular tutorial example. It
would also capture screenshots of the Blender window at appropriate
steps (likely every step) and save them along with annotation sufficient
to build a complete tutorial.

One of the goals of this project is to make it easier to keep our
tutorials up to date with the latest Blender and CellBlender.

This project is being started in the test suite section because the code
will be very similar to the test suite code. In fact, if this works
well, it might be combined with the test suite code to make tutorials
identical to what the test suite does. This would allow our tests and
our tutorials to work as either.