Implemented basic frame capture code in CellBlender's test suite.
CellBlender's test suite is already very close to what we need to
capture tutorials, so it makes sense to combine them to implement the
CellBlender portions of the project.

It also makes some sense to separate the capturing of images and the
conversion to HTML (or other formats). With that thought in mind, the
capture code generates a JSONish text file that can be easily read by
any other programs to target the tutorials to other platforms. We may
implement some of those as part of the tutorial capture, but they should
be able to run stand-alone as well.

So this commit implements the capture portion and leaves the
reformatting to other code.