Added code from tutorial_maker.py to generate the HTML in Blender.
This version includes 2 buttons:

   One to make the image frames
   One to make the HTML for the frames

This version will generate the image files and the HTML in the directory
where it is run from. This shouldn't be much of a problem since we will
generally be the ones running this (at least in the beginning).

This version creates a torus object, and then assigns it a series of
colors (r,g,b,y,m,c). It captures a frame at each color and puts each
frame into a PNG image (named "#_color_frame.png"), and it also creates
a file named "tutorial_list.txt" which is a mostly complete JSON file
with titles and descriptions. This list of frames and the text file
should be sufficient to generate a number of different output formats
(HTML of different flavors, PDF documents, etc). The second button in
this version generates an HTML file named "tutorial.html" to link the
frames together using CSS and Javascript.

To run this, start with an empty Blender scene and click the first
button. It should cycle through all of the frames at about 1 per second.
When it's done, click the second button which will generate the HTML for
all the frames. View the result by opening "tutorial.html".