Replaced entire data model with hard-coded fceri model (testing mode)
Work done on the bng/bngl_to_data_model.py converter:

This version attempts to parse the BNGL file (and does a reasonable job
for what's been implemented), but the model doesn't run with what's been
implemented so far. So after all the parsing of parameters and molecules,
this version eventually over-writes the entire data model with the fceri
model. This at least makes it run.

The approach to finish this is to compare the data model being built
with the hard-coded fceri model section by section. Each section should
be able to be matched ... eventually.