Replaced fully hard-coded version with partially hard-coded version
There may be a compatibility problem between BioNetGen expression of
compartments and MCells definition of release sites.

It appears that the CBNGL release expressions exclude internal
compartments, but this must be done explicitly in MCell. For example,
the fceri model releases Lig in the EC with this expression:

  @EC::Lig(l,l)  Lig_tot

However, the MCell equivalent appears to be:

  EC[ALL] - CP[ALL]

If these interpretations are correct, then the conversion program will
need to keep track of internal compartments and exclude them from the
resulting release expressions.

This commit doesn't attempt to address this problem and simply uses
hard-coded expressions from the fceri model.