Added proper support for molecule and quantity ... not obj_expr yet.
As mentioned in the previous commit, handling of the seed species
compartments will require dealing with the nesting of objects since
MCell includes contained objects within the release volume unless
explicitly excluded (maybe there's an MCell option for this?).

This could be a problem with observables as well, but the observables in
the example (fceri) don't seem to reference compartments in observables.