Added a compartment model compatible with fceri definitions
The fceri model uses EC and CP rather than ECF and CP1. But the
fceri model doesn't demonstrate the nested structures that are in
the BNGL Quick Reference Guide (QRG). This commit just added the
QRG compartments and renamed a subset of them to match the fceri
model so that the more complex geometry could be tested while still
running the fceri model.