Added a copy of bngl.ebnf from Tom Bartol's personal files.
This version of bngl.ebnf is from a local file with a date of:
  August 31st, 2017

It's not known if this was the original version or if it has
been superceded by subsequent versions. It is being added here
as a point in its history.

This may not be an authoritative version.