Added a tilde (~) prefix to "children" key to have it sort last (debug)
For debugging, this code uses the "dump_data_model" function to display
the nested list of compartments. The dump_data_model function sorts the
keys in each dictionary alphabetically. This is normally fine, except
that it prints the "children" of each object before most of the other
characteristics (like name, x, y, z, ...). This makes the dump much more
difficult to follow.

In order to have the "children" dumped last, this commit adds a tilde
prefix (~) to the key "children" to make that key: "~children". This
puts the other keys (like name, x, y, z, ...) first and the children
last. This was done to aid debugging, and the proper key of "children"
should be restored at a later date. Note that the dictionary using this
key is only used internally to this file so it shouldn't affect any
other code.

The following example shows the output with "~children" sorted last.
This makes the nesting much easier to follow.

Topology {}
  name = "World"
  x = 0
  xdim = 0.8000000000000003
  y = 0
  ydim = 0.7999999999999999
  z = 0
  zdim = 0.7999999999999999
  ~children {}
    EC {}
      dim = "3"
      name = "EC"
      vol = "vol_EC"
      x = -5.551115123125783e-17
      xdim = 0.7000000000000002
      y = 0
      ydim = 0.7
      z = 0
      zdim = 0.7
      ~children {}
        PM {}
          dim = "2"
          name = "PM"
          vol = "vol_PM"
          x = -1.1102230246251565e-16
          xdim = 0.6000000000000001
          y = 0
          ydim = 0.6
          z = 0
          zdim = 0.6
          ~children {}
            CP {}
              dim = "3"
              name = "CP"
              vol = "vol_CP"
              x = -1.6653345369377348e-16
              xdim = 0.5
              y = 0
              ydim = 0.5
              z = 0
              zdim = 0.5
              ~children {}