CBNGL to DM: Add volumes of inner compartments to parent volume
The volumes specified in the "compartments" section of a BNGL file do
not include the volumes of any contained compartments. So to generate
the properly sized cubes for singly nested compartments, the parent size
must be large enough to contain its own volume plus the volumes needed by
any contained compartments.

This commit adds the contained volumes to the parent volume before
computing the required dimensions (cube root of parent+child vols).

As with previous commits, this does not yet properly allocate volumes
when compartments are not singly nested.