Added parameter evaluation for computing volumes of compartments
Most of the parameters from a BNGL model can be stored in the data model
as strings. However, parameters may be used to define compartment
volumes which requires their evaluation when building the geometry.

This commit adds a lightweight evaluation of parameters assuming that
they comply with Python expression syntax and that they are defined
before they are used. This allows them to be evaluated easily with
Python's "eval" function when used to define compartment volumes.

Note that this simple-minded parsing assumes that the volume expression
can be isolated from the other parts of a compartment definition with
the Python "split" function. This requires that there be no embedded
spaces in the compartment names (probably OK), the compartment
dimensions (probably OK), and the volume expression itself (not sure).