Moved the molecules section down to get vol/surf types from earlier
A BNGL file doesn't explicitly define volume or surface molecules. They
must be implied from their release location (in a 2D or 3D compartment).

This information should be able to be collected during the processing of
compartments and seed species. In anticipation of collecting that data,
the molecules section has been moved down below the seed species
section in this commit.