Determine molecule type (2D/3D) from release context (seed species)
This commit builds a dictionary of object dimensionality keyed by the
object name. It then uses that object dimensionality dictionary to
build a dictionary of species dimensionalities based on the placement
(seed species) of each molecule type. All releases are not guaranteed
to be on objects of the same dimensionality, so this data is stored as
a list of dimensionalities (might be better as a set).

Eventually, when the molecules are created, they are created of the
first dimensionality found in their list (should we create two species
if the dimensionalities differ?). This provides a reasonably automatic
way to get the molecule type (2D or 3D) from a CBNGL file.

Note that the diffusion constant is still not known for any species.
This current version uses two fixed diffusion constants from the FceRI
model according to this code:

  if mol['mol_type'] == '3D':
    mol['diffusion_constant'] = "8.51e-7"
  else:
    mol['diffusion_constant'] = "1.7e-7"

A warning is also printed.