Deduce release sites from CBNGL seed species and compartments alone
This commit creates proper MCell release sites based on CBNGL seed
species. It subtracts child objects from the parent object since CBNGL
"compartments" do not include any contained volumes.

When the compartment of interest is a volume, the volume's children will
be surfaces. The volume's grandchildren will be volumes. In MCell
syntax, these grandchildren should be subtracted from the object of
interest since they are the volumes that should be removed to release in
the compartment that excludes them. This generates MDL such as:

  Comp[ALL] - GrandChild1[ALL] - GrandChild2[ALL] ...

When the compartment of interest is a CBNGL surface, that CBNGL surface
will be the thin volume surrounding the enclosed volume object. This is
a different viewpoint than used by MCell. MCell views the surface as
being part of the volume object rather than the volume object's parent.
Visually, CBNGL sees this as:

  --- surface(outer) ---
 |                      |
 |     volume(inner)    |
 |______________________|

where the volume is a nested child of the surface. But MCell sees the
surface as a part of the volume (almost as a child of the volume). So
MCell syntax for the surface would be:

  volume[surface]

But the same MCell expression from the CBNGL nesting viewpoint is:

  innner[outer]

This seems counter-intuitive at first, but it reflects the different
representations of objects and surfaces between the two programs.

Finally, this version still uses "hard-coded" values for volume and
surface diffusion constants (8.51e-7 and 1.7e-7 respectively). These
were the values from the FceRI example used as a template. It's not
clear where these values should be entered. The current CBNGL files
tend to use capitalized words for MCell parameters (like TIME_STEP,
ITERATIONS, and VACANCY_SEARCH_DISTANCE). We might use a convention
like "VOLUME_DIFFUSION_CONSTANT" and "SURFACE_DIFFUSION_CONSTANT" when
these can be specified. Otherwise, we might have to adopt a convention
for individual molecule names like "Lig_VOL_DIFFUSION_CONSTANT" and
"Lig_SURF_DIFFUSION_CONSTANT" (since a given molecule might take on the
role of both a volume molecule and a surface molecule).